lib: Convert internal sequence number to int64_t
With the usage of a 32 bit number as a integer, but storing
non-signed values in it, we have cases where numbers greater
than 2 billion are being read in and stored and used before
lower value numbers, which of course is awful and mean.

Fixes: #2126

+++ lib/plist_int.h
@@ -48,7 +48,7 @@ struct prefix_list {
 
 /* Each prefix-list's entry. */
 struct prefix_list_entry {
-	int seq;
+	int64_t seq;
 
 	int le;
 	int ge;
